+++ lib/slather/command/coverage_command.rb
@@ -24,6 +24,7 @@ class CoverageCommand < Clamp::Command
 
   option ["--input-format"], "INPUT_FORMAT", "Input format (gcov, profdata)"
   option ["--scheme"], "SCHEME", "The scheme for which the coverage was generated"
+  option ["--configuration"], "CONFIGURATION", "The configuration for test that the project was set"
   option ["--workspace"], "WORKSPACE", "The workspace that the project was built in"
   option ["--binary-file"], "BINARY_FILE", "The binary file against the which the coverage will be run", :multivalued => true
   option ["--binary-basename"], "BINARY_BASENAME", "Basename of the file against which the coverage will be run", :multivalued => true
@@ -42,6 +43,7 @@ class CoverageCommand < Clamp::Command
     setup_verbose_mode
     setup_input_format
     setup_scheme
+    setup_configuration
     setup_workspace
     setup_binary_file
     setup_binary_basename
@@ -127,6 +129,10 @@ class CoverageCommand < Clamp::Command
 
   def setup_scheme
     project.scheme = scheme
+  end
+
+  def setup_configuration
+    project.configuration = configuration
   end
 
   def setup_workspace